NETFLIX-BUILD: [BDC-1210] WAP gather step failing when INSERT OVERWRITE does not add any new data files and skips commit.
Merge in BDP/iceberg from BDC-1210-Commit-Snapshot-for-WAP-in-OverwriteDynamicMode to netflix/1.1.x

Squashed commit of the following:

commit 91c7e2d8afb5ddd5619a00fd4967af434844fbb7
Author: Dhruv Pratap <dpratap@netflix.com>
Date:   Fri Aug 25 13:34:47 2023 -0500

    BDC-1210: De-dup the check isWapEnabledForTableAndJob spread across SparkWriter.

commit 2aee52f0f64b6b526f255b93a7f380f780792066
Author: Dhruv Pratap <dpratap@netflix.com>
Date:   Thu Aug 24 17:12:27 2023 -0500

    BDC-1210: WAP gather step failing when INSERT OVERWRITE add any new data files and skips commit. WAP thus is unable to find a snapshot staged with its WAP ID and fails.

+++ core/src/main/java/org/apache/iceberg/BaseReplacePartitions.java
@@ -111,7 +111,8 @@ public class BaseReplacePartitions extends MergingSnapshotProducer<ReplacePartit
 
   @Override
   public List<ManifestFile> apply(TableMetadata base, Snapshot snapshot) {
-    if (dataSpec().fields().size() <= 0) {
+    // If there are no added files, this is a no-op
+    if (!addedDataFiles().isEmpty() && dataSpec().fields().size() <= 0) {
       // replace all data in an unpartitioned table
       deleteByRowFilter(Expressions.alwaysTrue());
     }
